+++ packages/rules-engine/src/defaults.ts
@@ -1,4 +1,3 @@
-import type { OperatorEvaluator } from 'json-rules-engine';
 import type { ClassnamesRE, RulesEngine, RulesEngineIC, TranslationsFullRE } from './types';
 
 /**
@@ -85,16 +84,4 @@ export const defaultTranslationsRE: TranslationsFullRE = {
     title: 'Evaluate every condition independently; any number may match',
     label: 'Evaluate all (cumulative)',
   },
-};
-
-export const jsonRulesEngineAdditionalOperators: Record<
-  string,
-  OperatorEvaluator<unknown, unknown>
-> = {
-  beginsWith: (factValue, compareToValue) => `${factValue}`.startsWith(`${compareToValue}`),
-  doesNotBeginWith: (factValue, compareToValue) => !`${factValue}`.startsWith(`${compareToValue}`),
-  endsWith: (factValue, compareToValue) => `${factValue}`.endsWith(`${compareToValue}`),
-  doesNotEndWith: (factValue, compareToValue) => !`${factValue}`.endsWith(`${compareToValue}`),
-  contains: (factValue, compareToValue) => `${factValue}`.includes(`${compareToValue}`),
-  doesNotContain: (factValue, compareToValue) => !`${factValue}`.includes(`${compareToValue}`),
 };